Refactoring Glamorous-ified Hero into Bulma components directory
